this enhancement tries to disambiguate memory access with complex address, Say access with address E1=*p+i*200+j*4 and E2=*p+(i+1)*200+(j+8)*4. In this case, no existing point-to analysis algorithm can help and LNO's dep-test may ignore them due to its complexity.
During the phase creating CODEREP, the IR is cannonicalized. In this case the E2 
is canonicalized into E1+232. So, when Add_idef() is called,  we interpret the  
iload/istore with addess E2 as a access from pointer E1 plus offset 232. We record
this information in the POINTS_TO structure, and hence render it possible to 
disambiguate accesses with address E1 and E2. 


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@782 ead328b1-e903-0410-a19d-eacee4b031f6