Merge the fix to bug 397 and 420: When initialize a array of struct, in original implementation, there is a temporary symbol used to construct the structs, then the temporary object is copied to destination bit-wised. In the new implementation, in order to avoid generating the temporary symbol, we pass the address of the distination to the Aggreagte_XXX functions to complete the construction. The second bit-wised copy is unnecessary at all.
There is a macro NEW_INITIALIZER in the patch to reserve original code. By 
default, this patch is disabled. If you want to enable it, please remove the
'#' before the 'HOSTDEFS += -DNEW_INITIALIZER' in osprey/wgen/Makefile.gbase and 
osprey/kg++fe/Makefile.gbase. This patch should be enabled and all original 
code should be removed later.



git-svn-id: http://svn.open64.net/svnroot/open64/trunk@1439 ead328b1-e903-0410-a19d-eacee4b031f6