This checkin adds VCG Graph Visualization infrastructure to the Open64 compiler.
VCG classes are provided to emit vcg graphs for any data structure.
These are available in common/util/vcg.h and vcg.cxx and are mostly
self explanatory.

A Whirl to VCG client is implemented as an example in 
be/com/com_whirlview.cxx.  It can be invoked from the 
debugger using vcg_wn(WN *wn, char *filename).   A vcg viewer will
be required to view the graph.  Current vcg client produces a 
hierarchical graph for a whirl tree where Blocks, If's, Loops, 
and any SCF nodes are subgraphs.  These can be folded and 
unfolded interactively by the viewer.  Currently any 
subgraph >= 25 nodes or edges is folded by default.  

Hope this will improve the developer experience with the 
Open64 compiler. 



git-svn-id: http://svn.open64.net/svnroot/open64/trunk@2788 ead328b1-e903-0410-a19d-eacee4b031f6