Bug 592: Assertion failure in file: osprey/../libspin/gspin-tel.h,          function: gs_tree_value, line: 947
It happens when the initializer list for vector typed variable is empty. 
The code used to get the element type from initializer, when it is empty, 
the code gets a null pointer access. Now it is changed to use the vector 
MTYPE to get its element type.
 
__attribute__ ((vector_size (sizeof (double) * 4))) double d1 = { 1., 2 };
__attribute__ ((vector_size (sizeof (double) * 4))) double d2 = { };
__attribute__ ((vector_size (sizeof (float) * 4))) float f1 = { 1., 2 };
__attribute__ ((vector_size (sizeof (float) * 4))) float f2 = { };
__attribute__ ((vector_size (sizeof (int) * 4))) int a = { 1, 2 };
__attribute__ ((vector_size (sizeof (int) * 4))) int c = { };

Code Review: JianXin Lai



git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3340 ead328b1-e903-0410-a19d-eacee4b031f6