Fix build failure with gcc-4.5.1 and recent changes to symtab_defs.h.
In trunk commit r3435 a reference to std::ostream was added to
symtab_defs.h.  With older versions of gcc, the include of <ext/slist>
in symtab.h indirectly included <iosfwd> which resolves the reference.
With more recent versions of gcc, <ext/slist> does not include <iosfwd>.
The fix is to explicitly include the file that is needed, <iosfwd>.

Approved by: Suneel Jain


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3438 ead328b1-e903-0410-a19d-eacee4b031f6