Fix for bug700.
The problem is in newly checked-in componentization code for processing the 
option (rev3390).

The logic to process the option ("-componentname:optionlist") was not exact. 
The change is made to check the first char to be "-" and then check the 
componentname by parsing ':' .

Another change is the place of processing options for the new 
componentization-based approach. Now it is inside Process_Command_Line().

Code Review: Sun Chan




git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3449 ead328b1-e903-0410-a19d-eacee4b031f6