Observation -----------         When multiple source files was provided in command line, open64 was not able to complete the compilation successfully.
Problem
-------
        The problem was, the temp files created using "mkstemp" was not
closed.

         As per the current implementation, we are only performing
unlink on the temp_files, in the cleanup() function.

        --cut--
        unlink() deletes a name from the filesystem. If that name was
the last link to a file and no processes have the file open the file is
deleted and the space it was using is made available for reuse. 
        If the name was the last link to a file but any processes still
have the file open the file will remain in existence until the last file
descriptor referring to it is closed. 
        --cut--
        
        Since we are not closing the opened temp files, unlink will not
be able to delete the temp files created.

Fix:
----
        The file descriptor (FD) of the temp files opened has to be
collected, so that in cleanup() function, we can close the opened files
before unlinking.

Implementation
--------------
        The existing "temp_files" list is of type string_list_t, which
has been mostly used for string related operations.

        So changed the "temp_files" type to  file_item_t (new data
structure) and all the necessary changes to perform cleanup has been
made.




git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3512 ead328b1-e903-0410-a19d-eacee4b031f6