Fix to bug 707.
delete the enhancement of dse which is added from open64-4.2.1, since
the enhancement is not aware of the fact that there is "implied mus" 
for all the variables that are aliased with the variable of the mu node. 
So the chi that is dead because there is no use of its result is not
really dead.
 



git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3519 ead328b1-e903-0410-a19d-eacee4b031f6