This patch fixes a "use before defined" problem exposed by building gamess with -Ofast -apo -IPA:noinline, which can cause a segmentation fault in the compiler during LNO.
The failure signature is that a segmentation fault occurs in
IPA_WN_MAP_Get() when Any_Loop_In_SNL_Parallelizable() is called
with a bad depth argument which was derived from an uninitialized
_depth field in an instance of class SNL_INFO.

Currently when FIZ_FUSE_INFO::Build() adds an element to the _sln_info
array (element type is SNL_INFO) in an instance of FIZ_FUSE_INFO, only
the _type field of the element was initialized.  With this change all
of the elements are initialized.

Reviewed by Sun Chan.


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3610 ead328b1-e903-0410-a19d-eacee4b031f6