Improve code alignment of basic blocks for x86 targets.
Use the "maximum number of bytes that should be skipped" parameter
to the .p2align assembler directive to provide better performance
for generated code.  If the branch-in frequency to the BB is higher,
then more padding bytes are allowed to align the BB, and vice-versa.
The exact values to use are determined empirically.

Approved by: Sun Chan


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3614 ead328b1-e903-0410-a19d-eacee4b031f6