Add experimental configure option to disable IPA support.
This option is useful for leaving out some of the less portable parts
of the compiler build.  Known portability problems specific to the IPA
support include:

o ELF objects are used as the container for the intermediate WHIRL.

o ipa_link uses dlopen and dlsym to directly call functions in ipa.so.

o IPA compilation assumes the presence of a POSIX make and shell.

Changes in this revision:

o Define make variable BUILD_SKIP_IPA when configured with --disable-ipa.
  In the top-level build, skip components that are specific to IPA when
  the variable is set to YES.

o In the driver, print a friendly error message if IPA compilation is
  attempted and BUILD_SKIP_IPA is defined.  Also leave out code related to
  IPA ELF objects (previously skipped when TARG_NVISA was defined).

Approved by: Mike Murphy


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3644 ead328b1-e903-0410-a19d-eacee4b031f6