Avoid unnecessary struct copies.
For the following code, the compiler generated unnecessary struct copies
for a return value that has a size too big to be passed in registers.

  typedef struct { char big[1024]; } C;
  C gc;
  extern C bar9 (void);
  extern C check9 (void) { return bar9 (); }

In this case there were two copies and one temporary variable generated.
After this change the compiler does not generate any extra copies.

The change handles the similar cases of initialization by function
call ("C c = bar9();") and assignment to a pointer ("*cp = bar9()").

For struct return values, the patch also includes a fix to follow the
x86-64 ABI: the caller provides space for the return value in a hidden
first argument and this address is returned in %rax or %eax for a
32-bit target.  Previously the compiler was not following this rule.

Approved by: Sun Chan


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3659 ead328b1-e903-0410-a19d-eacee4b031f6