Fix bug 584: overallocation of temp space (Fortran).
Consider the following test case:

  1 program test
  2   character (len=3233) :: lin
  3   character (len=1024) :: path, nam
  4   integer :: pos
  5   pos=3
  6   nam=path(1:(pos-1)) // lin
  7 end program

In order to store the result of the concatenation expression on line 7,
the Fortran front-end generates code to allocate a temporary buffer on
the stack.  In this case the size of the buffer should be calculated
as an addition of the two buffer sizes.  The bug was that instead of
allocating N bytes, N * N bytes were reserved.

This error would either be undetected, if there was sufficent stack space,
or fatal.  The test case shows an example that exceeds the 8192 kbytes
user stack limit common on modern Linux systems.

The fix is to pass the size of the element as the second parameter to
cwh_expr_temp in the Fortran front end, as documented in the function
definition.  We know it is the constant 1 because we have just created
a character array type.

Approved by: Michael Lai


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3665 ead328b1-e903-0410-a19d-eacee4b031f6