Testcase: extern double sqrt(double);
int global_arr[4];
void SetPoints()
{
  int i, r0;
  for (i=0; i<10; i++) {
    r0 = global_arr[3];
    sqrt(r0 * r0);
  }
}

Compiling with 'opencc -O3 -apo' gives:

### Assertion failure at line 710 of .../../../osprey/be/lno/access_vector.h:
### Compiler Error in file sample.c during Loop Nest Optimizer phase:
### SYMBOL::Init(WN*) called with opcode 66612
opencc INTERNAL ERROR: .../x86_64-open64-linux/4.2/be returned non-zero status 1

Problem/Fix Description:

The error is caused due to a dangling pointer leftover from an incomplete 
cleanup after array substitution removes a whirl node. The error shows up 
only for calls to builtin math function 'sqrt' (not sure if there are more 
such functions) where the original call is replaced by a builtin and a new 
call is placed depending on the accuracy of the result. The problem will not 
show with '-fno-math-errno'.
The fix adds a cleanup function to remove scalar references to WHIRL nodes 
marked for deletion.

The loop in sample.c:
  for (i=0; i<10; i++) {
    r0 = global_arr[3]; //L1
    sqrt(r0 * r0); //L2
  }

The CALL_INFO corresponding to L2 has 'r0' as a scalar_use.
Array substitution replaces 'r0' with 'global_arr[3] and deletes whirl node for 'r0'. 
But the reference to 'r0' still exists within call_info of 'sqrt', which
results in an inconsistency (assertion failure) in array_loop_info that is detected 
when merging a call's scalar uses as part auto parallelization.

Note that had the call to 'sqrt' been a a VCALL it would not go through array substitution.
'sqrt' being a builtin, the VCALL is replaced (in the absence of -fno-math-errno) by
F8SQRT, result saved in temp_var and depending on whether or not the result is a NAN,
a call is placed to F8CALL_SQRT. (See trace file with -Wb,-tra).
-apo is needed to expose error because Annotations of Call 
(NSE_Annotate_Scalar_Call called by IPA_LNO_Map_Calls) only occurs under autopar.

The patch to function 'DeleteElement' in  osprey/be/com/cxx_template.h is to avoid 
a compile error (during compiler build) caused by absence of a suitable assignment 
operator for SCALAR_NODE and SCALAR_REF. This is issue was not exposed earlier because 
existing calls to 'DeleteTop' are only on a stack of pointers. 

C.R by Sun Chan.


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3679 ead328b1-e903-0410-a19d-eacee4b031f6