For the ILOAD from address const 0xffffffffffff8001, the WHIRL IR is:   U8INTCONST -32767 (0xffffffffffff8001)  I4I4ILOAD 0 T<4,.predef_I4,4> T<53,anon_ptr.,8>
when creating the CODEREP, after calling CANON_CR::Trim_to_16bits(),
the base of ILOAD will be trunk to 0 and -32767 will be stored to the
offset of the ILOAD. In current WOPT, the generated IR is:
  U4INTCONST 0 (0x0)
 I4I4ILOAD -32767 T<4,.predef_I4,4> T<53,anon_ptr.,8>

We can find the type of the address is converted to U4. This is done
by CODEREP::Set_dtyp_const_val(MTYPE dt, INT64 v). When the constant
is in the range of U4, the dtyp will be set to U4. This will cause bad
code generation in CG. Since both the base and offset are 4 bytes, the
CG will generate ldc32 for the constant and finally generate the movl
instruction.

This patch will force the dtyp of address expression to be pointer type.

Code reviewed and approved by Sun Chan.


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3688 ead328b1-e903-0410-a19d-eacee4b031f6