Re-patch rev 3688 and re-fix bug 830.
Add_idef() is used for both ILOAD and ISTORE. For ISTORE, the lbase is NULL but previous patch doesn't check that. Also, previous patch doesn't work for store to constant address like:
*(int*)0xffffffffffff8001 = val;

The new patch checks if the lbase or sbase is NULL and works for both ILOAD/ISTORE to constant address.

Approved by Sun Chan.



git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3689 ead328b1-e903-0410-a19d-eacee4b031f6