Avoid creating unnecessary region exit blocks.
when control flow optimization is not being done, these BBs are not
needed and can increase memory use signficantly.

The flag OPT_Enable_EH_CFG_OPT controls whether the CFG_transformation
for EH regions is performed.  The default setting is to be enabled only
in mainopt.

Approved by: Mei Ye


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3696 ead328b1-e903-0410-a19d-eacee4b031f6