Enable more if-conversion in WOPT.
Convert:
if (a && b)
  x = ...
To:
if (a)
  if (b)
    x = ...

Without this transformation, WHIRL-lowering creates control flow
that WOPT-if-conversion can't handle.  With this transformation,
this now looks like any other if-conversion opportunity to WOPT.
So this transformation enables more if-conversion (i.e. replace
branch with conditional move).

The transformation is enabled when the internal flag
OPT_Lower_Splitsinglecand is on (the default) and when
if-conversion is enabled in WOPT.  It is applied when the action
flag LOWER_SHORTCIRCUIT is set (mainopt phase).

As part of the transformation, all profile info is updated to
be consistent: in-edge weights equal out-edge weights, all new
edges have weights, overall incoming/outgoing frequency to the
original node does not change, and the frequency of "if a && b"
is true is equal to the frequency of "if a/if b" is true.

Approved by: Sun Chan


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3697 ead328b1-e903-0410-a19d-eacee4b031f6