Fix EBO mul problem:
When compiling a function like:
long fooBig(long x) {
  return (x * (((long)1) << 63));
}

EBO will try to optimize this mul operation, in the function Convert_Imm_Mul. At the beginning of this function, the statement

INT64 val = imm_val < 0 ? -imm_val : imm_val;

is trying to get the absolute value of imm_val, but in this case, imm_val is ((long)1) << 63, which makes .imm_val is the same as imm_val. Then, in the loop

  int power = 0;
  while( val != 1 ){
    power++;
    val >>= 1;
  }

val will keep being -1, which make this loop endless.

The solution is to change the type of imm_val from INT64 to UINT64.

Approved by Sun Chan.


git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3701 ead328b1-e903-0410-a19d-eacee4b031f6