Fix for bug 815 https://bugs.open64.net/show_bug.cgi?id=815
code reviewed by Sun Chan
 
Tested the open64 regression suite with no new failures
I have tested with the h264 program as filed in the bug
report -- base/peak with ref data and other SPECCPU2006 (no new failures
with -O3 and -apo).
 
The problem here is specific to MP regions that contain somewhat
advanced reductions. 
 
The code that causes the assert is SYMBOL(WN_array_base(WN_kid0(wn)) in
"osprey/be/lno/minvariant.cxx". Here a symbol is requested but the
WN_opcode does not have the sym PROPERTY. The WN_opcode of array base is
actually an ADD. Similar issue is in line 4418 in
"osprey/be/com/wn_mp.cxx" when this reduction is localized. Here WN_st
is accessed on the address computing ADD node. The computed NULL value
causes a SEGV downstream in the call chain. To cause this error apply
the patch only to "osprey/be/lno/minvariant.cxx". My fix is to ignore
the reductions which do not have the sym property in the MP region. 
 
Here is the dump of reduction tree that causes the issues:
 
   U8U8LDID 0 <1,51,stats> T<57,anon_ptr.,8>
   U4INTCONST 400 (0x190)
  U8ADD
  U4INTCONST 10 (0xa)
   I4I4LDID 0 <2,8,_temp__ubtmp0> T<4,.predef_I4,4>
  U8I4CVT
 U8ARRAY 1 4
XPRAGMA 1 52 <null-st> # REDUCTION {line: 1/25}



git-svn-id: http://svn.open64.net/svnroot/open64/trunk@3767 ead328b1-e903-0410-a19d-eacee4b031f6